refactor: rank 패키지를 추가하여 관리 (#332)
